Port xdBot to GD 2.2081 / Geode 5.10.1
- mod.json: geode 5.10.1, gd 2.2081, v2.5.0, native keybind + button settings,
  optional Custom Keybinds dependency
- CMakeLists: cmake 3.25, C++23, drop custom setting source
- Replace deprecated Geode APIs (popups, settings events, CCARRAY_FOREACH,
  file::pick task -> async::spawn)
- Update 2.2081 signatures: processCommands, getModifiedDelta,
  dispatchKeyboardMSG, CheckpointObject::init
- Use GameToolbox fast_srand bindings instead of hardcoded seed address
- Docs: about.md, changelog.md, PORTING-2.2081.md; pin SDK in CI

+++ src/hacks/tps_bypass.cpp
@@ -32,7 +32,8 @@ class $modify(GJBaseGameLayer) {
         
     }
 
-    float getModifiedDelta(float dt) {
+    // 2.2081: getModifiedDelta returns a double
+    double getModifiedDelta(float dt) {
         if (!Global::get().tpsEnabled) return GJBaseGameLayer::getModifiedDelta(dt);
         if (Global::getTPS() == 240.f) return GJBaseGameLayer::getModifiedDelta(dt);
         if (!PlayLayer::get()) return GJBaseGameLayer::getModifiedDelta(dt);
